Separate inference files
